requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
mangum>=0.17.0
python-dotenv>=1.0.0
pydantic>=2.4.2
crewai>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic-settings>=2.0.0
gunicorn

